move uint module
