+++ crates/starknet-devnet-core/src/starknet/add_l1_handler_transaction.rs
@@ -54,9 +54,9 @@ mod tests {
     use starknet_types::contract_address::ContractAddress;
     use starknet_types::contract_class::{Cairo0ContractClass, ContractClass};
     use starknet_types::felt::Felt;
+    use starknet_types::rpc::state::Balance;
     use starknet_types::rpc::transactions::L1HandlerTransaction;
     use starknet_types::traits::HashProducer;
-    use starknet_types::uint::Balance;
 
     use crate::account::Account;
     use crate::constants::{
