+++ crates/starknet-devnet-core/src/system_contract.rs
@@ -2,7 +2,7 @@ use blockifier::state::state_api::StateReader;
 use starknet_types::contract_address::ContractAddress;
 use starknet_types::contract_class::{Cairo0Json, ContractClass};
 use starknet_types::felt::{ClassHash, Felt};
-use starknet_types::uint::Balance;
+use starknet_types::rpc::state::Balance;
 
 use crate::error::DevnetResult;
 use crate::state::state_readers::DictState;
