+++ crates/starknet-devnet-core/src/traits.rs
@@ -2,7 +2,7 @@ use blockifier::state::state_api::StateReader;
 use starknet_types::contract_address::ContractAddress;
 use starknet_types::contract_class::ContractClass;
 use starknet_types::felt::ClassHash;
-use starknet_types::uint::Balance;
+use starknet_types::rpc::state::Balance;
 
 use crate::account::FeeToken;
 use crate::error::DevnetResult;
